refactor(pi-codex-conversion): simplify code mode registration
