+++ packages/pi-codex-conversion/src/tools/code-mode/AGENTS.md
@@ -4,4 +4,5 @@
 - Keep the copied runtime behavior aligned deliberately, but conversion-specific activation and nested tools belong outside this directory.
 - Codex host source and protocol remain pinned under `code-mode/vendor/code-mode-src/`.
 - `host-client.ts` owns process/session transport; `host-protocol.ts` validates wire data; `delegate-runtime.ts` owns nested execution; `trace-*` owns bounded trace state.
-- `tools.ts` composes Pi registration, while `tool-result.ts`, `render-tracker.ts`, and `rendering.ts` own their respective output boundaries.
+- `tools.ts` is the registration entrypoint; `shared-runtime.ts`, `public-tools.ts`, and `tool-events.ts` own provider state, Pi tools, and hooks respectively.
+- `tool-result.ts`, `render-tracker.ts`, and `rendering.ts` own their respective output boundaries.
